Create Script4.sh
Command used to find the genes overlapping the DMRs. DMR file is output of Script3. The genes.bed is constructed using the gene annotation of NCBI and contains all genes and their start and end position. Output file contains chromosome, start and end position, Gene ID and Gene name for each gene found in a DMR.